Rewrite Day 11: count differently for better performance
I've been annoyed by this one running too slowly
for a week now.
The way to count has been changed: neighbours/visible
are counted first and then the results are used
instead of being computed as we go. This allows
a few optimisations such as checking if the
current cell is "#" before going further and
using the collection Counter.